Remove Spotify following features, simplify trending and profile pages
