+++ src/App.tsx
@@ -11,8 +11,8 @@ import { Artists } from "./components/Artists";
 import { Shows } from "./components/Shows";
 
 import { Trending } from "./components/Trending";
-import { EnhancedTrending } from "./components/EnhancedTrending";
-import { EnhancedProfilePage } from "./components/EnhancedProfilePage";
+import { SimpleTrending } from "./components/SimpleTrending";
+import { SimpleProfile } from "./components/SimpleProfile";
 
 import { AppLayout } from "./components/AppLayout";
 import { ScrollToTop } from "./components/ScrollToTop";
@@ -297,7 +297,7 @@ function App() {
         );
       case "trending":
         return (
-          <EnhancedTrending 
+          <SimpleTrending 
             onArtistClick={handleArtistClick}
             onShowClick={handleShowClick}
           />
@@ -327,7 +327,7 @@ function App() {
           );
         }
         return (
-          <EnhancedProfilePage 
+          <SimpleProfile 
             onArtistClick={handleArtistClick}
             onShowClick={handleShowClick}
           />
